requirements.txt
django==3.0.3
django-markdownx
flake8==3.4.1
django-extensions==1.9.9
djangorestframework==3.11.0
markdown==2.6.11
django-cors-headers
psycopg2==2.8.4
uWSGI==2.0.17